redirect to dashboard and made dashboard protected route
